# api/api.yaml
openapi: 3.0.0
info:
  title: Draw2Gather API
  version: 0.0.1
servers:
  - url: http://localhost:8080
    description: Test Server
  - url: https://api.draw2gather.online
    description: Production Server
tags:
  - name: User
  - name: Game
security:
  - cookieAuth: []
paths:
  /test:
    get:
      operationId: test_get
      tags:
        - User
      security: []
      parameters:
        - name: location
          in: query
          description: Location
          required: false
          style: form
          explode: false
          schema:
            type: array
            minItems: 2
            maxItems: 2
            items:
              type: number
              format: double
              minimum: -180
              maximum: 180
      responses:
        default:
          $ref: "#/components/responses/ErrorResponse"
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - name
                properties:
                  name:
                    type: string
    post:
      operationId: test
      tags:
        - User
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - files
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        default:
          $ref: "#/components/responses/ErrorResponse"
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - name
                properties:
                  name:
                    type: string
  /user:
    post:
      operationId: createUser
      security:
        - bearerAuth:
            - "role:admin"
        - cookieAuth: []
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
      responses:
        default:
          $ref: "#/components/responses/ErrorResponse"
        200:
          description: OK
    get:
      operationId: getUser
      tags:
        - User
      responses:
        default:
          $ref: "#/components/responses/ErrorResponse"
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - name
                properties:
                  id:
                    type: string
                  name:
                    type: string
  /games:
    post:
      operationId: createGame
      tags:
        - Game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - lang
                - max_players
                - visibility
              properties:
                lang:
                  $ref: "#/components/schemas/language"
                max_players:
                  type: integer
                  minimum: 4
                  maximum: 10
                visibility:
                  type: boolean
      responses:
        default:
          $ref: "#/components/responses/ErrorResponse"
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                properties:
                  id:
                    type: string
    get:
      operationId: getGames
      tags:
        - Game
      parameters:
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/language"
      responses:
        default:
          $ref: "#/components/responses/ErrorResponse"
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - offset
                  - limit
                  - total
                  - games
                properties:
                  offset:
                    type: integer
                  limit:
                    type: integer
                  total:
                    type: integer
                  games:
                    type: array
                    items:
                      type: object
                      required:
                        - id
                        - lang
                        - max_players
                        - current_players
                      properties:
                        id:
                          type: string
                        lang:
                          $ref: "#/components/schemas/language"
                        max_players:
                          type: integer
                        current_players:
                          type: integer
  /join:
    post:
      operationId: joinGame
      tags:
        - Game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  type: string
      responses:
        default:
          $ref: "#/components/responses/ErrorResponse"
        200:
          description: OK
components:
  schemas:
    language:
      type: string
      enum:
        - en
        - de
        - tr
  parameters:
    offset:
      name: offset
      in: query
      description: Number of items to skip
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    limit:
      name: limit
      in: query
      description: Max records to return
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 10
    language:
      name: lang
      in: query
      description: Language of the game
      required: true
      schema:
        $ref: "#/components/schemas/language"
  responses:
    ErrorResponse:
      description: Error Response
      content:
        application/json:
          schema:
            type: object
            required:
              - status
              - message
            properties:
              status:
                type: integer
              message:
                type: string
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: draw2gather
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT